+++ packages/component-library/draft/Kaizen/ZenNavigationBar/context.tsx
@@ -1,8 +1,0 @@
-import React from "react"
-import { NavigationChange } from "./types"
-
-type LinkClickContextType = {
-  handleNavigationChange?: NavigationChange
-}
-
-export const LinkClickContext = React.createContext<LinkClickContextType>({})